Modification produit ajout Article
les produits sont consideres comme les sous categories et les articles prenent la place de produits

+++ app/Http/Livewire/Stock/CategoriesTable.php
@@ -5,6 +5,7 @@ namespace App\Http\Livewire\Stock;
 use Livewire\Component;
 use App\Models\Product;
 use App\Models\Categorie;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Str;
 use Mediconesystems\LivewireDatatables\Column;
 use Mediconesystems\LivewireDatatables\NumberColumn;
@@ -53,22 +54,32 @@ class CategoriesTable extends LivewireDatatable
     }
     public function columns()
     {
-        return [
-            Column::name('name')
-                ->label('Name'),
+        if (Auth::user()->role == 'LOG1' || Auth::user()->role == 'ADMIN' || Auth::user()->role == 'Sup'){
+            return [
+                Column::name('name')
+                    ->label('Name'),
 
-            BooleanColumn::name('active')
-                ->label('State'),
+                BooleanColumn::name('active')
+                    ->label('State'),
 
-            Column::callback(['id','active'], function ($id,$active) {
+                Column::callback(['id','active'], function ($id,$active) {
 
-                $delete = '<a href="#" class="p-1 text-teal-600 hover:bg-teal-600  rounded" wire:click="deleteCategorie(' . $id . ')"><i class="icon-trash txt-danger"></i></a>';
-                $edit = '<a href="#" class="p-1 text-teal-600 hover:bg-teal-600 rounded" wire:click="edit(' . $id . ')" data-toggle="modal" data-target="#nCategorieModalForms"><i class="icon-pencil"></i></a>';
-                if ($active == false) {
-                    $delete = '<a href="#" class="p-1 text-teal-600 hover:bg-teal-600  rounded" wire:click="restoreCategorie(' . $id . ')"><i class="icon-action-undo txt-danger"></i></a>';
-                }
-                    return '<div class="flex space-x-1 justify-around">'.$edit . $delete .'</div>';
-            })->unsortable(),
-        ];
+                    $delete = '<a href="#" class="p-1 text-teal-600 hover:bg-teal-600  rounded" wire:click="deleteCategorie(' . $id . ')"><i class="icon-trash txt-danger"></i></a>';
+                    $edit = '<a href="#" class="p-1 text-teal-600 hover:bg-teal-600 rounded" wire:click="edit(' . $id . ')" data-toggle="modal" data-target="#nCategorieModalForms"><i class="icon-pencil"></i></a>';
+                    if ($active == false) {
+                        $delete = '<a href="#" class="p-1 text-teal-600 hover:bg-teal-600  rounded" wire:click="restoreCategorie(' . $id . ')"><i class="icon-action-undo txt-danger"></i></a>';
+                    }
+                        return '<div class="flex space-x-1 justify-around">'.$edit . $delete .'</div>';
+                })->unsortable(),
+            ];
+        }else {
+            return [
+                Column::name('name')
+                    ->label('Name'),
+
+                BooleanColumn::name('active')
+                    ->label('State'),
+            ];
+        }
     }
 }
